+++ libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: chanwoki <chanwoki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 14:41:54 by chanwoki          #+#    #+#             */
-/*   Updated: 2022/11/17 16:39:26 by chanwoki         ###   ########.fr       */
+/*   Updated: 2022/11/17 19:04:07 by chanwoki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strtrim(char const *s1, char const *set)
 {
-	int		start;
-	int		end;
+	size_t	start;
+	size_t	end;
 	char	*re;
 
 	if (s1 == 0)
